airflow/dags/requirements.txt
psycopg2-binary>=2.9,<3
sqlalchemy>=1.4,<2.0
apache-airflow-providers-postgres==5.6.0
awswrangler==3.12.0
boto3==1.38.35
botocore==1.38.35
certifi==2025.4.26
charset-normalizer==3.4.2
gspread==6.2.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
idna==3.10
jmespath==1.0.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pyarrow==20.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
